extract field fetching code to another class
